+++ r-containers.tf
@@ -1,5 +1,5 @@
 resource "azurerm_storage_container" "container" {
-  for_each = { for c in var.containers : c.name => c }
+  for_each = try({ for c in var.containers : c.name => c }, {})
 
   storage_account_name = azurerm_storage_account.storage.name