+++ r-tables.tf
@@ -1,5 +1,5 @@
 resource "azurerm_storage_table" "table" {
-  for_each = { for t in var.tables : t.name => t }
+  for_each = try({ for t in var.tables : t.name => t }, {})
 
   storage_account_name = azurerm_storage_account.storage.name